backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
fpdf2==2.7.7
google-generativeai==0.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
stripe==7.10.0
Pillow==10.2.0
PyPDF2==3.0.1
# Cloud deployment
psycopg2-binary==2.9.9
cloudinary==1.38.0
gunicorn==21.2.0
openai==1.55.3